requirements.txt
fastapi
uvicorn
python-multipart
pyzbar
pillow
httpx
pycryptodome
python-jose[cryptography]
bcrypt
aiofiles
hypothesis
pytest
pytest-asyncio
python-dotenv
